feat: Added support for utils package
The package.json file was updated to include the "@cocreate/utils" dependency. This change was required to implement some new functionality. In the src/index.js file, the implementation of the CoCreateAuthorize class was modified to add an object with two new properties: getValueFromObject and dotNotationToObject from "@cocreate/utils". Additional checks were also added to the checkArray function to delete keys from the data object and to include @cocreate/utils function when triggering CoCreateAuthorize.  A new function deleteKey was added into the code to remove specific keys from the data object.